+++ src/Addons/34thPRC_ArmourCustom/config.cpp
@@ -85,6 +85,7 @@ class CfgWeapons
 	class 34thPRC_ArmourStandard_ODST2_M52_Urban; //34thPRC_ArmourStandard
 	class 34thPRC_ArmourStandard_ODST2_M52_B_Urban; //34thPRC_ArmourStandard
 	class 34thPRC_ArmourStandard_ODST2_M52_C_Urban; //34thPRC_ArmourStandard
+	class 34thPRC_ArmourStandard_ODST2_M52_D_Urban; //34thPRC_ArmourStandard
 	class 34thPRC_ArmourStandard_ODST2_M52_Marksman_Urban; //34thPRC_ArmourStandard
 	class 34thPRC_ArmourStandard_ODST2_M52_CQB_Urban; //34thPRC_ArmourStandard