+++ src/changelog.md
@@ -6,6 +6,9 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## Unreleased
+### Added
+- New armor variant 'M52-D Combat Armor' for the ODSTs
+- Vasya custom armor
 
 ## 0.24.3
 ### Updated